refactor(validator): move tree advisor into owned tree slice root
