+++ calculogic-validator/doc/Audits/validator-shim-cleanup-design-checkpoint.md
@@ -19,15 +19,15 @@ Remaining shim-debt files and current canonical targets:
 1. `src/validator-runner.logic.mjs` → `src/core/validator-runner.logic.mjs`
 2. `src/validator-registry.knowledge.mjs` → `src/core/validator-registry.knowledge.mjs`
 3. `src/validator-scopes.runtime.mjs` → `src/core/validator-scopes.runtime.mjs`
-4. `src/tree-structure-advisor.host.mjs` → `src/tree/tree-structure-advisor.host.mjs`
-5. `src/tree-structure-advisor.logic.mjs` → `src/tree/tree-structure-advisor.logic.mjs`
+4. `src/tree-structure-advisor.host.mjs` → `tree/src/tree-structure-advisor.host.mjs`
+5. `src/tree-structure-advisor.logic.mjs` → `tree/src/tree-structure-advisor.logic.mjs`
 
 All five are single-line `export * from ...` forwarders with no independent runtime behavior.
 
 Boundary and policy observations relevant to these paths:
 
 - Canonical implementation ownership for runner/registry/scopes is already under `src/core/**`.
-- Canonical tree host ownership boundary is already under `src/tree/tree-structure-advisor.host.mjs`.
+- Canonical tree host ownership boundary is already under `tree/src/tree-structure-advisor.host.mjs`.
 - Public package exports currently expose `./runner`, `./registry`, and `./scopes` using the root shim files as export targets (subpath contract is public, concrete internal file target is implementation detail).
 - README currently includes an explicit compatibility note for legacy root tree advisor paths (`src/tree-structure-advisor.host.mjs`, `src/tree-structure-advisor.logic.mjs`).
 
@@ -134,7 +134,7 @@ Cleanup impact:
 Public/legacy expectation impact:
 
 - narrows compatibility promise to a clearly documented temporary tree legacy bridge
-- clarifies that canonical owned paths remain `src/core/**` and `src/tree/**`
+- clarifies that canonical owned paths remain `src/core/**` and `tree/src/**`
 
 Test strategy impact:
 
@@ -150,11 +150,11 @@ Why this is the best fit now:
 1. It improves ownership clarity by removing three root-core forwarders that no longer carry an explicit independent compatibility rationale.
 2. It preserves low-risk incrementalism by temporarily retaining the two tree root shims that currently have explicit compatibility wording in README.
 3. It avoids over-promising compatibility where policy intent is weak (core-root shims) while avoiding abrupt breakage where policy intent is currently explicit (tree-root shims).
-4. It keeps extraction paths obvious: canonical implementations remain in `src/core/**` and `src/tree/**`, and temporary legacy bridges are tightly scoped.
+4. It keeps extraction paths obvious: canonical implementations remain in `src/core/**` and `tree/src/**`, and temporary legacy bridges are tightly scoped.
 
 Policy statement (post-checkpoint target):
 
-- `src/core/**` and `src/tree/**` are canonical_source internal module ownership paths.
+- `src/core/**` and `tree/src/**` are canonical_source internal module ownership paths.
 - Root-level shim files are not a blanket compatibility surface.
 - Only explicitly documented temporary legacy paths should retain parity contracts, with a removal target declared in the next cleanup pass documentation.
 
@@ -177,7 +177,7 @@ Planned pass 3 actions:
 
 4. **Documentation alignment**
    - update cleanup pass 3 audit doc with explicit policy language distinguishing temporary legacy compatibility from canonical ownership
-   - keep public boundary carveouts unchanged (`src/index.mjs`, `src/naming/naming-validator.host.mjs`, `src/tree/tree-structure-advisor.host.mjs`)
+   - keep public boundary carveouts unchanged (`src/index.mjs`, `src/naming/naming-validator.host.mjs`, `tree/src/tree-structure-advisor.host.mjs`)
 
 Net expected result of pass 3 under this recommendation:
