+++ calculogic-validator/tree/src/tree-structure-advisor.contracts.mjs
