+++ calculogic-validator/tree/src/tree-structure-advisor.logic.mjs
